requirements.txt
requests-oauthlib ~= 1.3.0

responses ~= 0.10.14
pytest ~= 5.4.2
pytest-cov ~= 2.8.1
coveralls ~= 2.0.0
